.bashrc support's new LXD conventions
ie. no longer mount ~ directory, but instead just mount
or cp individual directories or files are required, so
that things like ~/.local don't get mixed up between host
and container.

This means that items such as 'back up bash-history' now
have conditionals to guard on whether destination dir
exists.

Also, superficially, remove some legacy logging from .profile,
used years ago to debug IBM RHEL running .profile more than
once per login.

+++ .bashrc
@@ -417,11 +417,15 @@ function historyc() {
 
 function history_backup () {
     (
-        # make a backup, overwriting other backups from today
-        cd ~/docs/config/bash_history/
-        \cp ~/.bash_history bash_history_$(date +%F)
-        # rm backups older than N days
-        ls -1 . | head -n -100 | xargs rm -f
+        if [ -d ~/docs/config/bash_history ] ; then
+            # make a backup, overwriting other backups from today
+            (
+            cd ~/docs/config/bash_history/
+            \cp ~/.bash_history bash_history_$(date +%F)
+            # rm backups older than N days
+            ls -1 . | head -n -100 | xargs rm -f
+            )
+        fi
     )
 }
 trap history_backup EXIT
